Moved the test resources to the __mocks__ folder
